+++ src/modules/navigator/components/file-browser/file-browser.vue
@@ -121,7 +121,7 @@ defineExpose({
     />
 
     <FileBrowserContent
-      :layout="layout"
+      :layout="props.layout"
     />
 
     <FileBrowserStatusBar